pyproject.toml
[project]
name = "seminar-paper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
    "bcrypt==3.2.2",
    "authlib>=1.6.0",
    "python-dotenv>=1.1.0",
    "httpx>=0.28.1",
    "starlette>=0.46.2",
    "itsdangerous>=2.2.0",
    "slowapi>=0.1.9",
    "ruff>=0.11.12",
    "requests>=2.32.3",
    "jwt>=1.3.1",
    "pytest>=8.4.1",
]
